fix texts
